Moved SelectableList into Menu as internal component
