+++ libs/@hashintel/ds-components/src/components/Menu/SelectableList/selectable-list-item.tsx
@@ -1,7 +1,7 @@
 import { cx } from "@hashintel/ds-helpers/css";
 
-import { Icon } from "../Icon/icon";
-import { LoadingSpinner } from "../Loading/loading-spinner";
+import { Icon } from "../../Icon/icon";
+import { LoadingSpinner } from "../../Loading/loading-spinner";
 import {
   checkIconSizeMap,
   arrowIconSizeMap,
@@ -9,7 +9,7 @@ import {
   indentUnitPx,
 } from "./selectable-list-item.recipe";
 
-import type { FormInputSize } from "../../util/form-shared";
+import type { FormInputSize } from "../../../util/form-shared";
 import type { ItemClasses } from "./selectable-list-item.recipe";
 import type { Item } from "./selectable-list-util";
