+++ libs/@hashintel/ds-components/src/components/Menu/menu.stories.tsx
@@ -1,15 +1,15 @@
 import { useCallback, useMemo, useState } from "react";
 
 import { Button } from "../Button/button";
-import { getItemId } from "../SelectableList/selectable-list-util";
+import { EllipsisMenu as EllipsisMenuComponent } from "./ellipsis-menu";
+import { Menu, type MenuItem } from "./menu";
+import { getItemId } from "./SelectableList/selectable-list-util";
 import {
   groupedItems,
   itemsWithSubActions,
-} from "../SelectableList/selectable-list.fixtures";
-import { EllipsisMenu as EllipsisMenuComponent } from "./ellipsis-menu";
-import { Menu, type MenuItem } from "./menu";
+} from "./SelectableList/selectable-list.fixtures";
 
-import type { Item, ItemOrGroup } from "../SelectableList/selectable-list";
+import type { Item, ItemOrGroup } from "./SelectableList/selectable-list";
 import type { Story, StoryDefault } from "@ladle/react";
 
 function prefixIds(
