+++ libs/@hashintel/ds-components/src/components/Menu/menu.tsx
@@ -3,17 +3,17 @@ import { Portal } from "@ark-ui/react/portal";
 import { cloneElement, useMemo } from "react";
 
 import { usePortalContainerRef } from "../../util/portal-container-context";
+import { type Position } from "../Tooltip/tooltip";
 import {
   SelectableList,
   type Item,
   type ItemOrGroup,
-} from "../SelectableList/selectable-list";
+} from "./SelectableList/selectable-list";
 import {
   getItemId,
   isGroup,
   useLoopSelection,
-} from "../SelectableList/selectable-list-util";
-import { type Position } from "../Tooltip/tooltip";
+} from "./SelectableList/selectable-list-util";
 
 export type MenuItem = Item & { selected?: boolean };
